ship the vendored cookbooks
While it is convenient to rely on the tools to do what we need for now
we need to vendor the cookbooks until the pantry cookbook itself is
public on Supermarket. If we don't vendor the cookbooks, the vendoring
step will fail on new systems that don't yet have `git` installed.

We can remove berks-cookbooks at a future date when we're using 100%
Supermarket-published cookbooks.